chore: [#43] python uv example working in compose.

+++ examples/__task__.py
@@ -1,3 +1,4 @@
+import json
 import os
 from __tasklib__ import TaskContext, TaskBuilder
 
@@ -5,24 +6,41 @@ mod = "compose"
 
 
 def _prefix(ctx: TaskContext):
+    # we are runnning at the root of the repo
+    conf_file = "config.env"
+    conf_arg = f"--env-file {conf_file}" if os.path.exists(conf_file) else ""
+    env_file = ".env"
+    env_arg = f"--env-file {env_file}" if os.path.exists(env_file) else ""
+    compose_file = os.path.join("examples", "docker-compose.yml")
+    service_profiles = os.getenv("SERVICE_PROFILES", "all")
+    profiles0 = [f"--profile {svc.strip()}" for svc in service_profiles.split()]
+    profiles = " ".join(profiles0)
     compose_file = os.path.join(ctx.project_dir, "docker-compose.yml")
-    return f"docker compose -f {compose_file} --profile all"
+    return f"""
+        docker compose \
+            -f {compose_file} \
+            {conf_arg} \
+            {env_arg} \
+            {profiles}
+        """
 
 
 def _up(ctx: TaskContext):
     ctx.log.info("Starting docker-compose")
-    ctx.exec(f"{_prefix(ctx)} up -d")
+    return ctx.exec(f"{_prefix(ctx)} up -d")
 
 
 def _down(ctx: TaskContext):
     ctx.log.info("Stopping docker-compose")
-    ctx.exec(f"{_prefix(ctx)} stop")
-    ctx.exec(f"{_prefix(ctx)} rm -f")
+    ret = ctx.exec(f"{_prefix(ctx)} stop")
+    if ret.returncode != 0:
+        return ret
+    return ctx.exec(f"{_prefix(ctx)} rm -f")
 
 
 def _logs(ctx: TaskContext):
     ctx.log.info("Showing docker-compose logs")
-    ctx.exec(f"{_prefix(ctx)} logs -f --tail 100")
+    return ctx.exec(f"{_prefix(ctx)} logs -f --tail 100")
 
 
 def _restart(ctx: TaskContext):
@@ -31,11 +49,41 @@ def _restart(ctx: TaskContext):
 
 
 def _pull(ctx: TaskContext):
-    ctx.exec(f"{_prefix(ctx)} pull")
+    return ctx.exec(f"{_prefix(ctx)} pull")
 
 
 def _log_service(ctx: TaskContext, service):
-    ctx.exec(f"{_prefix(ctx)} logs {service} -f --tail 100")
+    return ctx.exec(f"{_prefix(ctx)} logs {service} -f --tail 100")
+
+
+def _nuke(ctx: TaskContext):
+    ret = _down(ctx)
+    if ret.returncode != 0:
+        ctx.log.error(f"Failed to stop containers\nError was: {ret.stderr}")
+        return ret
+
+    # removes volumes so we can test from a clean state
+    compose_file = os.path.join(ctx.project_dir, "docker-compose.yml")
+    ret = ctx.exec(
+        f"docker compose -f {compose_file} --profile '*' config --format json",
+        capture=True,
+    )
+    if ret.returncode != 0:
+        ctx.log.error(f"Failed to get config\nError was: {ret.stderr}")
+        return ret
+
+    config = json.loads(ret.stdout)
+    project_name = config["name"]
+
+    ret = ctx.exec("docker volume ls", capture=True)
+    if ret.returncode != 0:
+        ctx.log.error(f"Failed to get config\nError was: {ret.stderr}")
+        return ret
+
+    for row in ret.stdout.split("\n"):
+        if project_name in row:
+            volume = row.split()[1]
+            ctx.exec(f"docker volume rm {volume}")
 
 
 def configure(builder: TaskBuilder):
@@ -45,3 +93,4 @@ def configure(builder: TaskBuilder):
     builder.add_task(mod, "ex:log", _logs)
     builder.add_task(mod, "ex:pull", _pull)
     builder.add_task(mod, "ex:log:caddy", lambda ctx: _log_service(ctx, "caddy"))
+    builder.add_task(mod, "ex:nuke", _nuke)
